Acceptance tests delete own  question/aswer spec

+++ app/controllers/questions_controller.rb
@@ -7,6 +7,8 @@ class QuestionsController < ApplicationController
   expose :question
 
   def create
+    question.user = current_user
+
     if question.save
       redirect_to question_path(question), notice: 'Your question was successfully created.'
     else
@@ -23,8 +25,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user&.author?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Your question was succesfully deleted.'
+    else
+      redirect_to question, alert: 'You are not the author of this question.'
+    end
   end
 
   private
